# api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: hello@permify.co
    name: API Support
    url: https://github.com/Permify/permify/issues
  description: Permify is an open source authorization service for creating fine-grained
    and scalable authorization systems.
  license:
    name: Apache-2.0 license
    url: https://github.com/Permify/permify/blob/master/LICENSE
  title: Permify API
  version: v0.9.9
servers:
- url: /
...
paths:
  /v1/tenants/create:
    post:
      operationId: tenants.create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreateRequest'
        description: TenantCreateRequest is the message used for the request to create
          a tenant.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantCreateResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: An unexpected error response.
      summary: create tenant
      tags:
      - Tenancy
      x-codeSamples:
      - label: go
        lang: go
        source: |-
          rr, err: = client.Tenancy.Create(context.Background(), &v1.TenantCreateRequest {
              Id: ""
              Name: ""
          })
      - label: node
        lang: javascript
        source: |-
          client.tenancy.create({
             id: "",
             name: ""
          }).then((response) => {
              // handle response
          })
      - label: cURL
        lang: curl
        source: |-
          curl --location --request POST 'http://localhost:3476/v1/tenants/create' \
          --header 'Content-Type: application/json' \
          --data-raw '{
              "id": "",
              "name": ""
          }'
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
  /v1/tenants/list:
    post:
      operationId: tenants.list
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantListRequest'
        description: TenantListRequest is the message used for the request to list
          all tenants.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantListResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: An unexpected error response.
      summary: list tenants
      tags:
      - Tenancy
      x-codeSamples:
      - label: go
        lang: go
        source: |-
          cr, err := client.Tenancy.List(context.Background(), &v1.TenantListRequest{
              PageSize: 20,
              ContinuousToken: "",
          })
      - label: node
        lang: javascript
        source: |-
          let res = client.tenancy.list({
              pageSize: 20,
              continuousToken: "",
          })
      - label: cURL
        lang: curl
        source: |-
          curl --location --request POST 'localhost:3476/v1/tenants/list' \
          --header 'Content-Type: application/json' \
          --data-raw '{
              "page_size": "10",
              "continuous_token": ""
          }'
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
  /v1/tenants/{id}:
    delete:
      operationId: tenants.delete
      parameters:
      - description: id is the unique identifier of the tenant to be deleted.
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantDeleteResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: An unexpected error response.
      summary: delete tenant
      tags:
      - Tenancy
      x-codeSamples:
      - label: go
        lang: go
        source: |-
          rr, err: = client.Tenancy.Delete(context.Background(), &v1.TenantDeleteRequest {
              Id: ""
          })
      - label: node
        lang: javascript
        source: |-
          client.tenancy.delete({
             id: "",
          }).then((response) => {
              // handle response
          })
      - label: cURL
        lang: curl
        source: curl --location --request DELETE 'http://localhost:3476/v1/tenants/t1'
      x-accepts:
      - application/json
...
components:
  schemas:
...
    Any:
      additionalProperties:
        type: object
      description: |-
        `Any` contains an arbitrary serialized protocol buffer message along with a
        URL that describes the type of the serialized message.

        Protobuf library provides support to pack/unpack Any values in the form
        of utility functions or additional generated methods of the Any type.

        Example 1: Pack and unpack a message in C++.

            Foo foo = ...;
            Any any;
            any.PackFrom(foo);
            ...
            if (any.UnpackTo(&foo)) {
              ...
            }

        Example 2: Pack and unpack a message in Java.

            Foo foo = ...;
            Any any = Any.pack(foo);
            ...
            if (any.is(Foo.class)) {
              foo = any.unpack(Foo.class);
            }
            // or ...
            if (any.isSameTypeAs(Foo.getDefaultInstance())) {
              foo = any.unpack(Foo.getDefaultInstance());
            }

         Example 3: Pack and unpack a message in Python.

            foo = Foo(...)
            any = Any()
            any.Pack(foo)
            ...
            if any.Is(Foo.DESCRIPTOR):
              any.Unpack(foo)
              ...

         Example 4: Pack and unpack a message in Go

             foo := &pb.Foo{...}
             any, err := anypb.New(foo)
             if err != nil {
               ...
             }
             ...
             foo := &pb.Foo{}
             if err := any.UnmarshalTo(foo); err != nil {
               ...
             }

        The pack methods provided by protobuf library will by default use
        'type.googleapis.com/full.type.name' as the type URL and the unpack
        methods only use the fully qualified type name after the last '/'
        in the type URL, for example "foo.bar.com/x/y.z" will yield type
        name "y.z".

        JSON
        ====
        The JSON representation of an `Any` value uses the regular
        representation of the deserialized, embedded message, with an
        additional field `@type` which contains the type URL. Example:

            package google.profile;
            message Person {
              string first_name = 1;
              string last_name = 2;
            }

            {
              "@type": "type.googleapis.com/google.profile.Person",
              "firstName": <string>,
              "lastName": <string>
            }

        If the embedded message type is well-known and has a custom JSON
        representation, that representation will be embedded adding a field
        `value` which holds the custom JSON in addition to the `@type`
        field. Example (for message [google.protobuf.Duration][]):

            {
              "@type": "type.googleapis.com/google.protobuf.Duration",
              "value": "1.212s"
            }
      example:
        '@type': '@type'
      properties:
        '@type':
          description: |-
            A URL/resource name that uniquely identifies the type of the serialized
            protocol buffer message. This string must contain at least
            one "/" character. The last segment of the URL's path must represent
            the fully qualified name of the type (as in
            `path/google.protobuf.Duration`). The name should be in a canonical form
            (e.g., leading "." is not accepted).

            In practice, teams usually precompile into the binary all types that they
            expect it to use in the context of Any. However, for URLs which use the
            scheme `http`, `https`, or no scheme, one can optionally set up a type
            server that maps type URLs to message definitions as follows:

            * If no scheme is provided, `https` is assumed.
            * An HTTP GET on the URL must yield a [google.protobuf.Type][]
              value in binary format, or produce an error.
            * Applications are allowed to cache lookup results based on the
              URL, or have them precompiled into a binary to avoid any
              lookup. Therefore, binary compatibility needs to be preserved
              on changes to types. (Use versioned type names to manage
              breaking changes.)

            Note: this functionality is not currently available in the official
            protobuf release, and it is not used for type URLs beginning with
            type.googleapis.com. As of May 2023, there are no widely used type server
            implementations and no plans to implement one.

            Schemes other than `http`, `https` (or the empty scheme) might be
            used with implementation specific semantics.
          type: string
      type: object
...
    Status:
      example:
        code: 0
        details:
        - '@type': '@type'
        - '@type': '@type'
        message: message
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
        details:
          items:
            $ref: '#/components/schemas/Any'
          type: array
      type: object
...
    Tenant:
      description: "Tenant represents a tenant with an id, a name, and a timestamp\
        \ indicating when it was created."
      example:
        name: name
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
      properties:
        id:
          description: The ID of the tenant.
          type: string
        name:
          description: The name of the tenant.
          type: string
        created_at:
          description: The time at which the tenant was created.
          format: date-time
          type: string
      type: object
    TenantCreateRequest:
      description: TenantCreateRequest is the message used for the request to create
        a tenant.
      example:
        name: name
        id: id
      properties:
        id:
          description: id is a unique identifier for the tenant.
          type: string
        name:
          description: name is the name of the tenant.
          type: string
      type: object
    TenantCreateResponse:
      description: TenantCreateResponse is the message returned from the request to
        create a tenant.
      example:
        tenant:
          name: name
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
      properties:
        tenant:
          $ref: '#/components/schemas/Tenant'
      type: object
    TenantDeleteResponse:
      description: TenantDeleteResponse is the message returned from the request to
        delete a tenant.
      example:
        tenant:
          name: name
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
      properties:
        tenant:
          $ref: '#/components/schemas/Tenant'
      type: object
    TenantListRequest:
      description: TenantListRequest is the message used for the request to list all
        tenants.
      example:
        continuous_token: continuous_token
        page_size: 0
      properties:
        page_size:
          description: |-
            page_size is the number of tenants to be returned in the response.
            The value should be between 1 and 100.
          format: int64
          type: integer
        continuous_token:
          description: |-
            continuous_token is an optional parameter used for pagination.
            It should be the value received in the previous response.
          type: string
      type: object
    TenantListResponse:
      description: TenantListResponse is the message returned from the request to
        list all tenants.
      example:
        tenants:
        - name: name
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
        - name: name
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
        continuous_token: continuous_token
      properties:
        tenants:
          description: tenants is a list of tenants.
          items:
            $ref: '#/components/schemas/Tenant'
          type: array
        continuous_token:
          description: continuous_token is a string that can be used to paginate and
            retrieve the next set of results.
          type: string
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey
...